modified to fix registration error

+++ server/server.js
@@ -24,13 +24,25 @@ app.post('/users', (req, res) => {
     if (err) {
       res.status(500).send('Error reading users file');
     } else {
-      const users = JSON.parse(data);
+      let users = [];
+      try {
+        users = JSON.parse(data);
+      } catch (parseErr) {
+        return res.status(500).send('Error parsing users file');
+      }
+
+      // Check if the user already exists
+      const userExists = users.some(user => user.username === newUser.username);
+      if (userExists) {
+        return res.status(400).send('User already exists');
+      }
+
       users.push(newUser);
       fs.writeFile('users.json', JSON.stringify(users), (err) => {
         if (err) {
           res.status(500).send('Error writing users file');
         } else {
-          res.status(201).send('User added');
+          res.status(201).json({ message: 'User added' });
         }
       });
     }
